test: consolidate skill attribution coverage
